Move autoGC() into GcStore
