+++ src/libstore/gc.cc
@@ -886,21 +886,25 @@ void LocalStore::collectGarbage(const GCOptions & options, GCResults & results)
 }
 
 
-void LocalStore::autoGC(bool sync)
+GcStore::~GcStore()
 {
-    static auto fakeFreeSpaceFile = getEnv("_NIX_TEST_FREE_SPACE_FILE");
+    std::shared_future<void> future;
 
-    auto getAvail = [this]() -> uint64_t {
-        if (fakeFreeSpaceFile)
-            return std::stoll(readFile(*fakeFreeSpaceFile));
+    {
+        auto state(_state.lock());
+        if (state->gcRunning)
+            future = state->gcFuture;
+    }
 
-        struct statvfs st;
-        if (statvfs(realStoreDir.get().c_str(), &st))
-            throw SysError("getting filesystem info about '%s'", realStoreDir);
+    if (future.valid()) {
+        printInfo("waiting for auto-GC to finish on exit...");
+        future.get();
+    }
+}
 
-        return (uint64_t) st.f_bavail * st.f_frsize;
-    };
 
+void GcStore::autoGC(bool sync)
+{
     std::shared_future<void> future;
 
     {
@@ -916,7 +920,7 @@ void LocalStore::autoGC(bool sync)
 
         if (now < state->lastGCCheck + std::chrono::seconds(settings.minFreeCheckInterval)) return;
 
-        auto avail = getAvail();
+        auto avail = getAvailableSpace();
 
         state->lastGCCheck = now;
 
@@ -929,7 +933,7 @@ void LocalStore::autoGC(bool sync)
         std::promise<void> promise;
         future = state->gcFuture = promise.get_future().share();
 
-        std::thread([promise{std::move(promise)}, this, avail, getAvail]() mutable {
+        std::thread([promise{std::move(promise)}, this, avail]() mutable {
 
             try {
 
@@ -950,7 +954,7 @@ void LocalStore::autoGC(bool sync)
 
                 collectGarbage(options, results);
 
-                _state.lock()->availAfterGC = getAvail();
+                _state.lock()->availAfterGC = getAvailableSpace();
 
             } catch (...) {
                 // FIXME: we could propagate the exception to the
