+++ src/libstore/local-store.cc
@@ -425,19 +425,6 @@ AutoCloseFD LocalStore::openGCLock()
 
 LocalStore::~LocalStore()
 {
-    std::shared_future<void> future;
-
-    {
-        auto state(_state.lock());
-        if (state->gcRunning)
-            future = state->gcFuture;
-    }
-
-    if (future.valid()) {
-        printInfo("waiting for auto-GC to finish on exit...");
-        future.get();
-    }
-
     try {
         auto fdTempRoots(_fdTempRoots.lock());
         if (*fdTempRoots) {
@@ -1950,5 +1937,18 @@ std::optional<std::string> LocalStore::getVersion()
     return nixVersion;
 }
 
+uint64_t LocalStore::getAvailableSpace()
+{
+    static auto fakeFreeSpaceFile = getEnv("_NIX_TEST_FREE_SPACE_FILE");
+
+    if (fakeFreeSpaceFile)
+        return std::stoll(readFile(*fakeFreeSpaceFile));
+
+    struct statvfs st;
+    if (statvfs(realStoreDir.get().c_str(), &st))
+        throw SysError("getting filesystem info about '%s'", realStoreDir);
+
+    return (uint64_t) st.f_bavail * st.f_frsize;
+}
 
 }  // namespace nix
